feat(vm): int is a signed 64-bit integer
An int held 32 bits, so a rowid past two billion, a unix time in
milliseconds, a file size, or 65536 * 65536 wrapped. It now holds the
signed 64-bit range from the literal through folding, the register, the
artifact, the host Value and the C boundary.

A tagged 64-bit word has no room for a 64-bit payload, so a value carries
the NaN box it always did plus a second word holding the int. Everything
else keeps its old encoding, and equality and hashing read one word for
every type but int. The .cdlb records the pair, which raises the format
version to 7. The SetInt operand stays narrow and a wider constant moves
out of its register, so an instruction is still one word. An index is
checked against the length in 64 bits before it becomes a usize, so a
position past 2^31 is out of bounds rather than a truncated read.

+++ vm/src/errors.rs
@@ -315,9 +315,9 @@ pub enum ErrType<'a> {
     // BOOL PARSING ERRORS
     InvalidBool,
     /// IndexOutOfBounds(length, index)
-    IndexOutOfBounds(usize, i32),
+    IndexOutOfBounds(usize, i64),
     /// SliceOutOfBounds(length, idx_start, idx_end)
-    SliceOutOfBounds(usize, i32, i32),
+    SliceOutOfBounds(usize, i64, i64),
     UnknownMapKey(&'a str),
     NullByteInString,
     CArrayReturnTypeNotSupported,
@@ -326,7 +326,7 @@ pub enum ErrType<'a> {
     ModuloByZero,
     /// An `int` raised to a negative `int` power, which has no `int` result.
     /// The parser rejects the all-literal form; this is the run-time one.
-    NegativeExponent(i32),
+    NegativeExponent(i64),
     /// A json string could not be parsed. Carries a short static reason.
     JsonParse(&'static str),
     /// A downcast of an `any` value found a different runtime type than the
@@ -345,7 +345,7 @@ pub enum ErrType<'a> {
     /// A byte position that lands in a character taking more than one byte,
     /// where a whole character is the only answer. Carries the position and
     /// the character it landed in.
-    NotACharBoundary(i32, char),
+    NotACharBoundary(i64, char),
 }
 
 impl From<ErrType<'_>> for SmolStr {
